added "perms" flag
Added the "perms" flag to the bindfs mount so that the permissions from the host are not only used, but **shown** within the VM (e.g. when doing a `ls -la` command. PHP (and surely other things as well) check these permissions to see if the directory/file is writeable (http://php.net/manual/en/function.is-writable.php) or not. If it **doesn't** appear writeable, PHP will error out (even if it actually is).